+++ ruby/CHANGELOG.md
@@ -6,6 +6,7 @@ This is the changelog for the `essence` gem. For now, most of the updates will b
 
 - feature: Adding `Card` component
 - feature: Adding `Checkbox` component
+- feature: Adding `Input` component
 - feature: Adding `Label` component
 - feature: Adding `Separator` component
 - feature: Adding `Switch` component